pyproject.toml
[tool.poetry]
name = "viziogram"
version = "0.1.0"
description = ""
authors = ["JoseRobertsVery <jose.r@verypossible.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask-RESTful = "^0.3.9"
Flask-SQLAlchemy = "^2.5.1"
flask-marshmallow = "^0.14.0"
flask-swagger-ui = "^4.11.1"
marshmallow-sqlalchemy = "^0.28.1"
Flask-Migrate = "^3.1.0"
flask-apispec = "^0.11.4"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
pytest = "^7.1.2"
pytest-watch = "^4.2.0"
autopep8 = "^1.7.0"
pre-commit = "^2.20.0"
coverage = "^6.4.4"
pylint = "^2.15.0"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
